@@ -1,6 +1,6 @@
 {
   "name": "react-native-slider",
-  "version": "0.6.1",
+  "version": "0.11.0",
   "description": "A pure JavaScript <Slider /> component for react-native",
   "main": "lib/Slider.js",
   "files": [
@@ -27,17 +27,15 @@
     "url": "git@github.com:jeanregisser/react-native-slider.git"
   },
   "dependencies": {
-    "react-addons-shallow-compare": "^0.14.8",
-    "style-equal": "^1.0.0"
+    "prop-types": "^15.5.6"
   },
   "devDependencies": {
     "babel-cli": "^6.6.5",
-    "babel-eslint": "^3.1.15",
-    "babel-preset-react-native": "^1.5.6",
-    "eslint": "^0.23.0",
-    "eslint-plugin-react": "^2.5.2",
-    "react": "^0.14.8",
-    "react-native": "^0.24.1",
+    "babel-eslint": "^7.2.1",
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "^6.10.3",
+    "react": "16.0.0-alpha.6",
+    "react-native": "0.43.3",
     "rimraf": "^2.5.2"
   }
 }
